// yangisdev/src/types/projects.ts
export type Project = {
    id: string
    projectField?: string
    name?: string
    logo?: string
    shortDescription?: string
    description?: string
    thumbnail?: string
    slug?: string
    screenshots?: string[]
    ref?: string
    technologies?: string[]
    content?: string
    startedAt?: string
    completedAt?: string
    createdAt?: string
    updatedAt?: string
}